This should fix the announcement system, among other changes
Notably, adding the announcement modifier option, which allows for
changing when the announcement would be posted relative to the event and
its other rules. Right now not supported on the front end at all, but
the backend is prepped for its support with the announcement posting
system soon to follow.