This should in theory fix the bad auth when getting events by API
Google is returning a 400 code saying we are using a bad key. Hopefully by using the same methods that are already working elsewhere in the software (specifically the abstracted reactive wrappers I made), this should at least help narrow down where/what is causing the issue or outright solve the issue.

This marks a point where I realize I need to go through the API and make sure its more in line with how other parts of the program gets the same data.