Removing more shared resources
Still seeing some issues with auth, not nearly as many as before, but still more than zero which is where it needs to be at like it was in v3. So here's hoping this helps since I'm suspecting its somehow one way or another getting the wrong auth info sometimes? Not sure...